requirements.txt
boto3>=1.37.24
psycopg2-binary>=2.9.0
